refactor: add PolicyConditionDistance
